Added RangeSlider and RangeSliderDual components. Refactored FilterContent component range slider into RangeSliderDual.
